server/requirements.txt
flask >= 0.10
sqlalchemy >= 0.8.2
flask-sqlalchemy >= 1.0
pyyaml >= 3.10
markdown >= 2.3.1
psycopg2 >= 2.5.1
